requirements.txt
fastmcp>=2.0.0
pymupdf>=1.24.0
python-docx>=0.8.11
mobi>=0.3.0
pytesseract>=0.3.10
docx2pdf>=0.1.8
python-dotenv>=1.0.0
Pillow>=10.0.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.7.0
numpy>=1.24.0
langdetect>=1.0.9